Déplacement de partie GUI dans le script session
