+++ __init__.py
@@ -1,3 +1,5 @@
+from .sammo import Sammo
+
+
 def classFactory(iface):
-    from .sammo import Sammo
     return Sammo(iface)
