+++ sammo.py
@@ -1,19 +1,13 @@
-from PyQt5.QtWidgets import QAction, QMessageBox
+from .src.gui.session import SammoActionSession
 
 
 class Sammo:
     def __init__(self, iface):
         self.iface = iface
-        self.action = None
+        self.actionSession = SammoActionSession(iface)
 
     def initGui(self):
-        self.action = QAction('Go!', self.iface.mainWindow())
-        self.action.triggered.connect(self.run)
-        self.iface.addToolBarIcon(self.action)
+        self.actionSession.initGui()
 
     def unload(self):
-        self.iface.removeToolBarIcon(self.action)
-        del self.action
-
-    def run(self):
-        QMessageBox.information(None, 'Minimal plugin', 'Do something useful here')
+        self.actionSession.unload()
